+++ headers/private/kernel/vm/vm.h
@@ -176,7 +176,7 @@ status_t _user_get_memory_properties(team_id teamID, const void *address,
 area_id _user_area_for(void *address);
 area_id _user_find_area(const char *name);
 status_t _user_get_area_info(area_id area, area_info *info);
-status_t _user_get_next_area_info(team_id team, int32 *cookie, area_info *info);
+status_t _user_get_next_area_info(team_id team, ssize_t *cookie, area_info *info);
 status_t _user_resize_area(area_id area, size_t newSize);
 area_id _user_transfer_area(area_id area, void **_address, uint32 addressSpec,
 			team_id target);
